test_agents_structure.py
"""
Unit test script to verify all 7 agents load correctly without API calls.

This test validates code structure, imports, and configuration without
requiring Gemini API quota.
"""

import os
import sys
from dotenv import load_dotenv
import inspect

# Load environment variables
load_dotenv()

# Add src to path
sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'src'))

from utils.config import Settings
from agents.input_parser import InputParserAgent, ParsedInput
from agents.intent_detector import IntentDetectorAgent, EmailIntent
from agents.draft_writer import DraftWriterAgent
from agents.tone_stylist import ToneStylistAgent
from agents.personalization import PersonalizationAgent
from agents.review_agent import ReviewAgent
from agents.router import RouterAgent
from langchain_google_genai import ChatGoogleGenerativeAI


def print_section(title):
    """Print a formatted section title"""
    print(f"\n{'='*70}")
    print(f"  {title}")
    print(f"{'='*70}\n")


def test_config():
    """Test configuration loading"""
    print_section("TEST 1: Configuration Loading")
    try:
        config = Settings()
        print(f"✅ Config loaded successfully")
        print(f"   - App Name: {config.app_name}")
        print(f"   - Debug: {config.debug}")
        print(f"   - Model: {config.gemini_model}")
        print(f"   - API Key present: {'Yes' if config.gemini_api_key else 'No'}")
        print(f"   - Temperature: {config.temperature}")
        print(f"   - Max Tokens: {config.max_tokens}")
        return True
    except Exception as e:
        print(f"❌ Config loading failed: {e}")
        return False


def test_llm_initialization():
    """Test LLM initialization (not calling API)"""
    print_section("TEST 2: LLM Initialization")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key,
            temperature=config.temperature,
            max_tokens=config.max_tokens
        )
        print(f"✅ LLM initialized successfully")
        print(f"   - Type: {type(llm).__name__}")
        print(f"   - Model: {config.gemini_model}")
        return llm
    except Exception as e:
        print(f"❌ LLM initialization failed: {e}")
        return None


def test_agent_structure(agent_class, agent_name):
    """Test agent class structure"""
    try:
        # Check class exists and has required methods
        required_methods = ['__call__']
        methods = [m[0] for m in inspect.getmembers(agent_class, predicate=inspect.ismethod)]
        
        has_call = hasattr(agent_class, '__call__')
        has_init = hasattr(agent_class, '__init__')
        
        status = "✅" if (has_init and has_call) else "⚠️ "
        print(f"{status} {agent_name}")
        print(f"   - __init__: {has_init}")
        print(f"   - __call__: {has_call}")
        
        return True
    except Exception as e:
        print(f"❌ {agent_name} structure test failed: {e}")
        return False


def test_input_parser_structure():
    """Test Input Parser Agent structure"""
    print_section("TEST 3: Input Parser Agent")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key
        )
        parser = InputParserAgent(llm)
        
        print(f"✅ InputParserAgent instantiated")
        print(f"   - Has __call__ method: {hasattr(parser, '__call__')}")
        print(f"   - Has parse method: {hasattr(parser, 'parse')}")
        print(f"   - Prompt template: {type(parser.prompt).__name__}")
        
        # Check ParsedInput model
        print(f"   - ParsedInput fields: {list(ParsedInput.model_fields.keys())}")
        return True
    except Exception as e:
        print(f"❌ InputParserAgent test failed: {e}")
        return False


def test_intent_detector_structure():
    """Test Intent Detector Agent structure"""
    print_section("TEST 4: Intent Detector Agent")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key
        )
        detector = IntentDetectorAgent(llm)
        
        print(f"✅ IntentDetectorAgent instantiated")
        print(f"   - Has detect method: {hasattr(detector, 'detect')}")
        print(f"   - Has __call__ method: {hasattr(detector, '__call__')}")
        
        # Check intent enum
        intents = [intent.value for intent in EmailIntent]
        print(f"   - Supported intents: {len(intents)}")
        print(f"   - Intent types: {', '.join(intents[:3])}...")
        
        return True
    except Exception as e:
        print(f"❌ IntentDetectorAgent test failed: {e}")
        return False


def test_draft_writer_structure():
    """Test Draft Writer Agent structure"""
    print_section("TEST 5: Draft Writer Agent")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key
        )
        writer = DraftWriterAgent(llm)
        
        print(f"✅ DraftWriterAgent instantiated")
        print(f"   - Has write method: {hasattr(writer, 'write')}")
        print(f"   - Has __call__ method: {hasattr(writer, '__call__')}")
        print(f"   - Has prompt templates: {hasattr(writer, 'prompt')}")
        
        return True
    except Exception as e:
        print(f"❌ DraftWriterAgent test failed: {e}")
        return False


def test_tone_stylist_structure():
    """Test Tone Stylist Agent structure"""
    print_section("TEST 6: Tone Stylist Agent")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key
        )
        stylist = ToneStylistAgent(llm)
        
        print(f"✅ ToneStylistAgent instantiated")
        print(f"   - Has adjust_tone method: {hasattr(stylist, 'adjust_tone')}")
        print(f"   - Has __call__ method: {hasattr(stylist, '__call__')}")
        print(f"   - Tone guidelines count: {len(stylist.TONE_GUIDELINES)}")
        print(f"   - Available tones: {list(stylist.TONE_GUIDELINES.keys())}")
        
        return True
    except Exception as e:
        print(f"❌ ToneStylistAgent test failed: {e}")
        return False


def test_personalization_structure():
    """Test Personalization Agent structure"""
    print_section("TEST 7: Personalization Agent")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key
        )
        personalizer = PersonalizationAgent(llm)
        
        print(f"✅ PersonalizationAgent instantiated")
        print(f"   - Has personalize method: {hasattr(personalizer, 'personalize')}")
        print(f"   - Has __call__ method: {hasattr(personalizer, '__call__')}")
        print(f"   - Has get_profile method: {hasattr(personalizer, 'get_profile')}")
        print(f"   - Has save_profile method: {hasattr(personalizer, 'save_profile')}")
        
        return True
    except Exception as e:
        print(f"❌ PersonalizationAgent test failed: {e}")
        return False


def test_review_agent_structure():
    """Test Review Agent structure"""
    print_section("TEST 8: Review Agent")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key
        )
        reviewer = ReviewAgent(llm)
        
        print(f"✅ ReviewAgent instantiated")
        print(f"   - Has review method: {hasattr(reviewer, 'review')}")
        print(f"   - Has __call__ method: {hasattr(reviewer, '__call__')}")
        
        return True
    except Exception as e:
        print(f"❌ ReviewAgent test failed: {e}")
        return False


def test_router_agent_structure():
    """Test Router Agent structure"""
    print_section("TEST 9: Router Agent")
    try:
        config = Settings()
        llm = ChatGoogleGenerativeAI(
            model=config.gemini_model,
            api_key=config.gemini_api_key
        )
        router = RouterAgent(llm)
        
        print(f"✅ RouterAgent instantiated")
        print(f"   - Has route_next_step method: {hasattr(router, 'route_next_step')}")
        print(f"   - Has __call__ method: {hasattr(router, '__call__')}")
        
        return True
    except Exception as e:
        print(f"❌ RouterAgent test failed: {e}")
        return False


def test_imports():
    """Test all imports work correctly"""
    print_section("TEST 10: All Imports Verification")
    try:
        # These are already imported at the top, but let's verify
        print(f"✅ All imports successful")
        print(f"   - Settings: {Settings.__name__}")
        print(f"   - InputParserAgent: {InputParserAgent.__name__}")
        print(f"   - IntentDetectorAgent: {IntentDetectorAgent.__name__}")
        print(f"   - DraftWriterAgent: {DraftWriterAgent.__name__}")
        print(f"   - ToneStylistAgent: {ToneStylistAgent.__name__}")
        print(f"   - PersonalizationAgent: {PersonalizationAgent.__name__}")
        print(f"   - ReviewAgent: {ReviewAgent.__name__}")
        print(f"   - RouterAgent: {RouterAgent.__name__}")
        print(f"   - ChatGoogleGenerativeAI: {ChatGoogleGenerativeAI.__name__}")
        
        return True
    except Exception as e:
        print(f"❌ Import verification failed: {e}")
        return False


def main():
    """Run all structural tests"""
    print("\n" + "="*70)
    print("  EMAIL GENERATOR APP - STRUCTURAL VALIDATION TEST SUITE")
    print("="*70)
    print("  (No API calls - validates code structure and imports)")
    
    results = []
    
    # Run tests
    results.append(("Configuration Loading", test_config()))
    results.append(("LLM Initialization", test_llm_initialization() is not None))
    results.append(("Input Parser Structure", test_input_parser_structure()))
    results.append(("Intent Detector Structure", test_intent_detector_structure()))
    results.append(("Draft Writer Structure", test_draft_writer_structure()))
    results.append(("Tone Stylist Structure", test_tone_stylist_structure()))
    results.append(("Personalization Structure", test_personalization_structure()))
    results.append(("Review Agent Structure", test_review_agent_structure()))
    results.append(("Router Agent Structure", test_router_agent_structure()))
    results.append(("All Imports", test_imports()))
    
    # Summary
    print_section("TEST SUMMARY")
    passed = sum(1 for _, result in results if result)
    total = len(results)
    
    print(f"Results: {passed}/{total} tests passed\n")
    for name, result in results:
        status = "✅ PASS" if result else "❌ FAIL"
        print(f"{status}: {name}")
    
    print("\n" + "="*70)
    if passed == total:
        print("✅ ALL STRUCTURAL TESTS PASSED!")
        print("\nYour code is ready for the next phase:")
        print("  • All 7 agents are properly structured")
        print("  • All imports work correctly")
        print("  • Configuration is loaded successfully")
        print("  • LLM is initialized and authenticated")
        print("\nNote: Free tier quota exceeded for Gemini API")
        print("Next steps:")
        print("  1. Check your Gemini API quota at: https://ai.google.dev/pricing")
        print("  2. Consider upgrading to paid tier or wait for quota reset")
        print("  3. Implement LangGraph workflow (src/workflow/langgraph_flow.py)")
        print("  4. Build Streamlit UI (src/ui/streamlit_app.py)")
    else:
        print("⚠️  SOME TESTS FAILED - Please review the errors above")
    print("="*70 + "\n")
    
    return passed == total


if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)
